+++ test/core/store/stores/test_dir.py
@@ -1,10 +1,14 @@
 import os.path
 import unittest
 
+from xcube.core.store.store import new_data_store
 from xcube.core.store.stores.dir import DirectoryDataStore
 
 
 class DirectoryCubeStoreTest(unittest.TestCase):
+    def test_new_data_store(self):
+        store = new_data_store('dir', base_dir='.')
+        self.assertIsInstance(store, DirectoryDataStore)
 
     def test_get_data_ids(self):
         data_store = DirectoryDataStore(
