+++ xcube/cli/_gen2/write.py
@@ -18,12 +18,12 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-import xarray as xr
 from typing import Callable, Optional
 
+import xarray as xr
+
 from xcube.cli._gen2.request import OutputConfig
-from xcube.core.store.helpers import new_cube_store
-from xcube.core.store.store import CubeWriter
+from xcube.core.store.store import new_data_store
 from xcube.util.extension import ExtensionRegistry
 
 
@@ -31,12 +31,11 @@ def write_cube(cube: xr.Dataset,
                output_config: OutputConfig,
                progress_monitor: Callable,
                extension_registry: Optional[ExtensionRegistry] = None):
-    cube_store = new_cube_store(output_config.cube_store_id,
-                                output_config.cube_store_params,
-                                cube_store_requirements=(CubeWriter,),
-                                extension_registry=extension_registry)
+    data_store = new_data_store(output_config.cube_store_id,
+                                extension_registry=extension_registry,
+                                **output_config.cube_store_params)
     cube_id = output_config.cube_id
-    write_params_schema = cube_store.get_write_cube_params_schema()
+    write_params_schema = data_store.get_write_data_params_schema()
     write_params = write_params_schema.from_json(output_config.write_params) \
         if output_config.write_params else {}
-    cube_store.write_cube(cube, cube_id=cube_id, write_params=write_params)
+    data_store.write_cube(cube, data_id=cube_id, **write_params)
