+++ xcube/core/store/stores/dir.py
@@ -23,20 +23,27 @@ import os.path
 import uuid
 from typing import Optional, Iterator, Any, Tuple, List
 
+import geopandas as gpd
 import xarray as xr
 
 from xcube.core.mldataset import MultiLevelDataset
-from xcube.core.store.accessor_v4 import get_data_opener
-from xcube.core.store.accessor_v4 import get_data_opener_extensions
-from xcube.core.store.accessor_v4 import get_data_writer
-from xcube.core.store.accessor_v4 import get_data_writer_extensions
-from xcube.core.store.descriptor import DataDescriptor, get_data_type_id, TYPE_ID_DATASET, TYPE_ID_MULTI_LEVEL_DATASET, \
-    TYPE_ID_GEO_DATA_FRAME
+from xcube.core.store.accessor import find_data_opener_extensions
+from xcube.core.store.accessor import find_data_writer_extensions
+from xcube.core.store.accessor import get_data_accessor_predicate
+from xcube.core.store.accessor import new_data_opener
+from xcube.core.store.accessor import new_data_writer
+from xcube.core.store.descriptor import DataDescriptor
 from xcube.core.store.descriptor import DatasetDescriptor
 from xcube.core.store.descriptor import MultiLevelDatasetDescriptor
-from xcube.core.store.store_v4 import DataStoreError
-from xcube.core.store.store_v4 import MutableDataStore
-from xcube.util.assertions import assert_in, assert_instance, assert_given
+from xcube.core.store.descriptor import TYPE_ID_DATASET
+from xcube.core.store.descriptor import TYPE_ID_GEO_DATA_FRAME
+from xcube.core.store.descriptor import TYPE_ID_MULTI_LEVEL_DATASET
+from xcube.core.store.descriptor import get_data_type_id
+from xcube.core.store.store import DataStoreError
+from xcube.core.store.store import MutableDataStore
+from xcube.util.assertions import assert_given
+from xcube.util.assertions import assert_in
+from xcube.util.assertions import assert_instance
 from xcube.util.extension import Extension
 from xcube.util.jsonschema import JsonBooleanSchema
 from xcube.util.jsonschema import JsonObjectSchema
@@ -115,12 +122,14 @@ class DirectoryDataStore(MutableDataStore):
         self._assert_valid_type_id(type_id)
         if not type_id and data_id:
             type_id, _, _ = self._get_accessor_id_parts(data_id)
-        return tuple(ext.name for ext in get_data_opener_extensions(type_id=type_id, storage_id='posix'))
+        return tuple(ext.name for ext in find_data_opener_extensions(
+            predicate=get_data_accessor_predicate(type_id=type_id, storage_id='posix')
+        ))
 
     def get_open_data_params_schema(self, data_id: str = None, opener_id: str = None) -> JsonObjectSchema:
         if not opener_id and data_id:
             opener_id = self._get_opener_id(data_id)
-        return get_data_opener(opener_id).get_open_data_params_schema(data_id=data_id)
+        return new_data_opener(opener_id).get_open_data_params_schema(data_id=data_id)
 
     def open_data(self,
                   data_id: str,
@@ -129,17 +138,22 @@ class DirectoryDataStore(MutableDataStore):
         if not opener_id:
             opener_id = self._get_opener_id(data_id)
         path = self._resolve_data_id_to_path(data_id)
-        return get_data_opener(opener_id).open_data(path, **open_params)
+        return new_data_opener(opener_id).open_data(path, **open_params)
 
     def get_data_writer_ids(self, type_id: str = None) -> Tuple[str, ...]:
         self._assert_valid_type_id(type_id)
-        return tuple(ext.name for ext in get_data_writer_extensions(type_id=type_id, storage_id='posix'))
+        extensions = find_data_writer_extensions(
+            predicate=get_data_accessor_predicate(type_id=type_id, storage_id='posix')
+        )
+        return tuple(ext.name for ext in extensions)
 
     def get_write_data_params_schema(self, writer_id: str = None) -> JsonObjectSchema:
         if not writer_id:
-            extensions = get_data_writer_extensions(type_id='dataset', storage_id='posix')
+            extensions = find_data_writer_extensions(
+                predicate=get_data_accessor_predicate(type_id='dataset', storage_id='posix')
+            )
             writer_id = extensions[0].name
-        return get_data_writer(writer_id).get_write_data_params_schema()
+        return new_data_writer(writer_id).get_write_data_params_schema()
 
     def write_data(self,
                    data: Any,
@@ -147,23 +161,34 @@ class DirectoryDataStore(MutableDataStore):
                    writer_id: str = None,
                    replace: bool = False,
                    **write_params) -> str:
-        assert_instance(data, (xr.Dataset, MultiLevelDataset))
+        assert_instance(data, (xr.Dataset, MultiLevelDataset, gpd.GeoDataFrame))
         if not writer_id:
             if isinstance(data, xr.Dataset):
-                extensions = get_data_writer_extensions(type_id='dataset', format_id='zarr', storage_id='posix')
+                predicate = get_data_accessor_predicate(type_id=TYPE_ID_DATASET,
+                                                        format_id='zarr',
+                                                        storage_id='posix')
+            elif isinstance(data, MultiLevelDataset):
+                predicate = get_data_accessor_predicate(type_id=TYPE_ID_MULTI_LEVEL_DATASET,
+                                                        format_id='levels',
+                                                        storage_id='posix')
+            elif isinstance(data, gpd.GeoDataFrame):
+                predicate = get_data_accessor_predicate(type_id=TYPE_ID_GEO_DATA_FRAME,
+                                                        format_id='geojson',
+                                                        storage_id='posix')
             else:
-                extensions = get_data_writer_extensions(type_id='mldataset', format_id='levels', storage_id='posix')
+                raise DataStoreError(f'Unsupported data type {type(data)}')
+            extensions = find_data_writer_extensions(predicate=predicate)
             writer_id = extensions[0].name
         data_id = self._ensure_valid_data_id(data_id, data)
         path = self._resolve_data_id_to_path(data_id)
-        get_data_writer(writer_id).write_data(data, path, replace=replace, **write_params)
+        new_data_writer(writer_id).write_data(data, path, replace=replace, **write_params)
         return data_id
 
     def delete_data(self, data_id: str):
         accessor_id_parts = self._get_accessor_id_parts(data_id)
         writer_id = ':'.join(accessor_id_parts)
         path = self._resolve_data_id_to_path(data_id)
-        get_data_writer(writer_id).delete_data(path)
+        new_data_writer(writer_id).delete_data(path)
 
     def register_data(self, data_id: str, data: Any):
         # We don't need this as we use the file system
@@ -189,16 +214,16 @@ class DirectoryDataStore(MutableDataStore):
             assert_in(type_id, self.get_type_ids(), 'type_id')
 
     def _get_opener_id(self, data_id: str):
-        return self._get_accessor_id(data_id, get_data_opener_extensions)
+        return self._get_accessor_id(data_id, find_data_opener_extensions)
 
     def _get_writer_id(self, data_id: str):
-        return self._get_accessor_id(data_id, get_data_writer_extensions)
+        return self._get_accessor_id(data_id, find_data_writer_extensions)
 
     def _get_opener_extensions(self, data_id: str, require=True):
-        return self._get_accessor_extensions(data_id, get_data_opener_extensions, require=require)
+        return self._get_accessor_extensions(data_id, find_data_opener_extensions, require=require)
 
     def _get_writer_extensions(self, data_id: str, require=True):
-        return self._get_accessor_extensions(data_id, get_data_writer_extensions, require=require)
+        return self._get_accessor_extensions(data_id, find_data_writer_extensions, require=require)
 
     def _get_accessor_id(self, data_id: str, get_data_accessor_extensions, require=True) -> Optional[str]:
         extensions = self._get_accessor_extensions(data_id, get_data_accessor_extensions, require=require)
@@ -231,4 +256,3 @@ class DirectoryDataStore(MutableDataStore):
     def _get_filename_ext(cls, data: Any):
         type_id = get_data_type_id(data)
         return _TYPE_ID_TO_ACCESSOR_TO_DEFAULT_EXT[type_id]
-
